Split cann_accl into accel/ subpackage
- new canns/models/basic/accel/ package with four internal modules:
  _protocol.py (IrecBackend), _backends.py (4 backend classes),
  _spectrum.py (_pick_k_for_err_target), _factory.py
  (make_irec_backend + pre-compute helpers). Public surface
  re-exported from accel/__init__.py with the same module
  docstring (decision table + 'how to add a backend' guide).
- cann.py: from .cann_accl → from .accel. All cross-references
  in docstrings updated. Public surface unchanged
  (ACCL_MODES, ACCL_DEFAULT_K, _pick_k_for_err_target still
  importable from canns.models.basic.cann).
- 09_acceleration_modes.ipynb (en + zh): cross-refs updated,
  re-executed against the new layout. All 11 code cells
  still produce identical outputs.
